zmon-scheduler/requirements.txt
dogpile.cache
redis
suds
graphitesend
PyYAML
requests
eventlog-writer==0.4.5
dnspython
cherrypy>=3.0.0
netaddr

#zalando-cmdb-client==0.9.1
#zalando-deployctl-client==0.2
#zalando-zompy==0.2.17
#zalando-scmmanager==0.0.23
